Yoni's progress (#3)
* Adding class CrashedBroadcast by ass - vol 3.1
etc in whatsapp and discord

* LiDarService & LiDarWorkerTracker has to be written again

* maybe now u can see?

* Basic implementation of all objects

---------

+++ src/main/java/bgu/spl/mics/application/messages/broadcasts/TickBroadcast.java
@@ -3,13 +3,13 @@ package bgu.spl.mics.application.messages.broadcasts;
 import bgu.spl.mics.Broadcast;
 
 public class TickBroadcast implements Broadcast {
-    private final long tick;
+    private final int tick;
 
-    public TickBroadcast(long tick) {
+    public TickBroadcast(int tick) {
         this.tick = tick;
     }
 
-    public long getTick() {
+    public int getTick() {
         return tick;
     }
 }
